Split case participant nodes into subpackage
- replace monolithic case participant nodes module with focused
  participant/ submodules for helpers, owner flow, participant add flow,
  and status snapshot flow
- preserve legacy import compatibility at
  vultron.core.behaviors.case.nodes.participant via package re-exports
  for public nodes and helper symbols
- mirror the split in tests by replacing test_participant.py with
  focused participant/ test modules and shared fixtures
