--- vultron/core/behaviors/case/nodes/participant.py
@@ -1,1209 +1,0 @@
-#!/usr/bin/env python
-
-#  Copyright (c) 2026 Carnegie Mellon University and Contributors.
-#  - see Contributors.md for a full list of Contributors
-#  - see ContributionInstructions.md for information on how you can Contribute to this project
-#  Vultron Multiparty Coordinated Vulnerability Disclosure Protocol Prototype is
-#  licensed under a MIT (SEI)-style license, please see LICENSE.md distributed
-#  with this Software or contact permission@sei.cmu.edu for full terms.
-#  Created, in part, with funding and support from the United States Government
-#  (see Acknowledgments file). This program may include and/or can make use of
-#  certain third party source code, object code, documentation and other files
-#  ("Third Party Software"). See LICENSE.md for more details.
-#  Carnegie Mellon®, CERT® and CERT Coordination Center® are registered in the
-#  U.S. Patent and Trademark Office by Carnegie Mellon University
-
-"""
-Participant management action nodes and helpers for case behavior trees.
-
-Provides helpers and action nodes that create and attach case participants,
-manage participant statuses, and emit participant-add notifications.
-
-Per specs/case-management.yaml CM-02-008, BTND-05-002.
-"""
-
-import logging
-from typing import TYPE_CHECKING, Any, cast
-
-import py_trees
-from py_trees.common import Status
-
-from vultron.core.behaviors.helpers import DataLayerAction
-from vultron.core.models.actor_config import ActorConfig
-from vultron.core.models.case_status import CaseStatus
-from vultron.core.models.enums import VultronObjectType
-from vultron.core.models.participant_status import ParticipantStatus
-from vultron.core.models.protocols import CaseModel, is_case_model
-from vultron.core.models.vultron_types import VultronCase, VultronParticipant
-from vultron.core.ports.case_persistence import (
-    CaseOutboxPersistence,
-    CasePersistence,
-)
-from vultron.core.states.cs import CS_pxa, CS_vfd
-from vultron.core.states.em import EM
-from vultron.core.states.participant_embargo_consent import PEC
-from vultron.core.states.rm import RM
-from vultron.core.states.roles import CVDRole
-from vultron.core.use_cases._helpers import (
-    _as_id,
-    _report_phase_status_id,
-    update_participant_rm_state,
-)
-
-if TYPE_CHECKING:
-    from vultron.core.ports.trigger_activity import TriggerActivityPort
-
-logger = logging.getLogger(__name__)
-
-
-# ============================================================================
-# Shared helper functions
-# ============================================================================
-
-
-def _create_and_attach_participant(
-    dl: CasePersistence,
-    participant: "VultronParticipant",
-    case_id: str,
-    actor_id_for_index: str,
-    node_logger: logging.Logger,
-) -> "CaseModel | None":
-    """
-    Create a VultronParticipant in the DataLayer if it does not exist,
-    attach it to the case, and return the **unsaved** updated case object.
-
-    The caller is responsible for any additional case mutations (e.g.,
-    ``record_event``, RM advancement) and for calling ``dl.save(case)``
-    after this function returns.
-
-    Args:
-        dl: DataLayer instance for persistence.
-        participant: The participant object to create and attach.
-        case_id: ID of the case to attach the participant to.
-        actor_id_for_index: Actor ID key for the case's
-            ``actor_participant_index``.
-        node_logger: Logger from the calling BT node.
-
-    Returns:
-        The updated (unsaved) case object on success, or ``None`` if the
-        case was not found.
-    """
-    if dl.read(participant.id_) is None:
-        dl.create(participant)
-        node_logger.info(
-            "Created CaseParticipant '%s' for actor '%s'",
-            participant.id_,
-            participant.attributed_to,
-        )
-    else:
-        node_logger.debug(
-            "CaseParticipant %s already exists — skipping creation",
-            participant.id_,
-        )
-
-    stored_case = dl.read(case_id)
-    if not is_case_model(stored_case):
-        node_logger.error("Case %s not found in DataLayer", case_id)
-        return None
-
-    existing_participant_id = stored_case.actor_participant_index.get(
-        actor_id_for_index
-    )
-    if existing_participant_id is not None:
-        existing_participant = dl.read(existing_participant_id)
-        if existing_participant is not None:
-            stored_case.add_participant(existing_participant)
-            node_logger.debug(
-                "Participant already registered for actor '%s' in case '%s'",
-                actor_id_for_index,
-                case_id,
-            )
-            return stored_case
-
-    stored_case.add_participant(participant)
-
-    node_logger.info(
-        "CaseParticipant '%s' attached to case '%s'",
-        participant.id_,
-        stored_case.id_,
-    )
-    return stored_case
-
-
-def _resolve_case_id(
-    blackboard: Any, case_obj: VultronCase | None = None
-) -> str | None:
-    case_id = case_obj.id_ if case_obj is not None else None
-    return case_id or blackboard.get("case_id")
-
-
-def _build_owner_initial_status(
-    dl: CasePersistence,
-    actor_id: str,
-    case_id: str,
-    report_id: str | None,
-    initial_rm_state: RM,
-) -> ParticipantStatus:
-    if report_id is not None:
-        status_id = _report_phase_status_id(
-            actor_id,
-            report_id,
-            initial_rm_state.value,
-        )
-        if dl.read(status_id) is not None:
-            return ParticipantStatus(
-                id_=status_id,
-                context=case_id,
-                rm_state=initial_rm_state,
-                attributed_to=actor_id,
-            )
-
-    return ParticipantStatus(
-        context=case_id,
-        rm_state=initial_rm_state,
-        attributed_to=actor_id,
-    )
-
-
-def _effective_case_roles(actor_config: ActorConfig | None) -> list[CVDRole]:
-    base_roles = actor_config.default_case_roles if actor_config else []
-    return list(dict.fromkeys(base_roles + [CVDRole.CASE_OWNER]))
-
-
-def _get_or_create_accepted_status(
-    dl: CasePersistence,
-    actor_id: str,
-    report_id: str | None,
-    node_name: str,
-    node_logger: logging.Logger,
-) -> ParticipantStatus | None:
-    if report_id is None:
-        return None
-
-    accepted_status_id = _report_phase_status_id(
-        actor_id,
-        report_id,
-        RM.ACCEPTED.value,
-    )
-    existing = dl.read(accepted_status_id)
-    if isinstance(existing, ParticipantStatus):
-        return existing
-
-    node_logger.info(
-        "%s: Creating fresh RM.ACCEPTED status for actor '%s' "
-        "(report-phase status not pre-created)",
-        node_name,
-        actor_id,
-    )
-    accepted_status = ParticipantStatus(
-        id_=accepted_status_id,
-        context=report_id,
-        rm_state=RM.ACCEPTED,
-        attributed_to=actor_id,
-    )
-    try:
-        dl.create(accepted_status)
-    except ValueError:
-        pass
-    return accepted_status
-
-
-def resolve_participant_state_from_dl(
-    dl: CasePersistence,
-    participant_id: str,
-) -> tuple[RM, CS_vfd]:
-    """Return (current_rm, current_vfd) from the participant's latest status.
-
-    Reads the participant record from the DataLayer and extracts the RM
-    and VFD state values from the most recent entry in
-    ``participant_statuses``.  Falls back to ``(RM.START, CS_vfd.vfd)``
-    when the participant is not found or has no statuses.
-
-    Args:
-        dl: DataLayer instance.
-        participant_id: ID of the CaseParticipant to inspect.
-
-    Returns:
-        ``(rm_state, vfd_state)`` tuple with validated enum values.
-    """
-    participant_obj = dl.read(participant_id)
-    if participant_obj is not None and hasattr(
-        participant_obj, "participant_statuses"
-    ):
-        statuses = getattr(participant_obj, "participant_statuses")
-        if statuses:
-            latest = statuses[-1]
-            raw_rm = getattr(latest, "rm_state", RM.START)
-            raw_vfd = getattr(latest, "vfd_state", CS_vfd.vfd)
-            rm_state = raw_rm if isinstance(raw_rm, RM) else RM.START
-            vfd_state = raw_vfd if isinstance(raw_vfd, CS_vfd) else CS_vfd.vfd
-            return rm_state, vfd_state
-    return RM.START, CS_vfd.vfd
-
-
-def _queue_participant_add_notification(
-    dl: CasePersistence,
-    node_name: str,
-    node_logger: logging.Logger,
-    sender_actor_id: str,
-    participant_actor_id: str,
-    participant_id: str,
-    case_id: str,
-    trigger_activity: "TriggerActivityPort | None" = None,
-) -> bool:
-    stored_participant = dl.read(participant_id)
-    if (
-        getattr(stored_participant, "type_", None)
-        != VultronObjectType.CASE_PARTICIPANT
-    ):
-        node_logger.error(
-            "%s: Could not resolve stored CaseParticipant '%s'",
-            node_name,
-            participant_id,
-        )
-        return False
-
-    if trigger_activity is None:
-        node_logger.error(
-            "%s: trigger_activity_factory not available for participant"
-            " add notification",
-            node_name,
-        )
-        return False
-
-    add_notification_id = trigger_activity.add_participant_to_case(
-        participant_id=participant_id,
-        case_id=case_id,
-        actor=sender_actor_id,
-        to=[participant_actor_id],
-    )
-
-    cast(CaseOutboxPersistence, dl).record_outbox_item(
-        sender_actor_id, add_notification_id
-    )
-    node_logger.info(
-        "Queued Add(CaseParticipant '%s' for actor '%s' to case '%s') "
-        "activity '%s' to actor '%s' outbox",
-        participant_id,
-        participant_actor_id,
-        case_id,
-        add_notification_id,
-        sender_actor_id,
-    )
-    return True
-
-
-# ============================================================================
-# Action Nodes
-# ============================================================================
-
-
-class ResolveOwnerInitialStatusNode(DataLayerAction):
-    """Resolve/create the owner's initial ParticipantStatus."""
-
-    def __init__(
-        self,
-        report_id: str | None,
-        case_obj: VultronCase | None,
-        initial_rm_state: RM,
-        name: str | None = None,
-    ) -> None:
-        super().__init__(name=name or self.__class__.__name__)
-        self.report_id = report_id
-        self.case_obj = case_obj
-        self.initial_rm_state = initial_rm_state
-
-    def setup(self, **kwargs: Any) -> None:
-        super().setup(**kwargs)
-        self.blackboard.register_key(
-            key="case_id", access=py_trees.common.Access.READ
-        )
-        self.blackboard.register_key(
-            key="owner_initial_status", access=py_trees.common.Access.WRITE
-        )
-
-    def update(self) -> Status:
-        if self.datalayer is None or self.actor_id is None:
-            self.logger.error(
-                "%s: DataLayer or actor_id not available",
-                self.name,
-            )
-            return Status.FAILURE
-
-        try:
-            case_id = _resolve_case_id(self.blackboard, self.case_obj)
-        except KeyError:
-            case_id = None
-        if case_id is None:
-            self.logger.error("%s: case_id not available", self.name)
-            return Status.FAILURE
-
-        self.blackboard.owner_initial_status = _build_owner_initial_status(
-            self.datalayer,
-            self.actor_id,
-            case_id,
-            self.report_id,
-            self.initial_rm_state,
-        )
-        return Status.SUCCESS
-
-
-class CreateOwnerParticipantNode(DataLayerAction):
-    """Create the in-memory owner participant and stage it on blackboard."""
-
-    def __init__(
-        self,
-        actor_config: ActorConfig | None,
-        name: str | None = None,
-    ) -> None:
-        super().__init__(name=name or self.__class__.__name__)
-        self.actor_config = actor_config
-
-    def setup(self, **kwargs: Any) -> None:
-        super().setup(**kwargs)
-        self.blackboard.register_key(
-            key="case_id", access=py_trees.common.Access.READ
-        )
-        self.blackboard.register_key(
-            key="owner_initial_status", access=py_trees.common.Access.READ
-        )
-        self.blackboard.register_key(
-            key="new_case_participant",
-            access=py_trees.common.Access.WRITE,
-        )
-
-    def update(self) -> Status:
-        if self.actor_id is None:
-            self.logger.error("%s: actor_id not available", self.name)
-            return Status.FAILURE
-        try:
-            case_id_obj = self.blackboard.get("case_id")
-        except KeyError:
-            case_id_obj = None
-        initial_status = self.blackboard.get("owner_initial_status")
-        if not isinstance(initial_status, ParticipantStatus):
-            self.logger.error(
-                "%s: case_id/owner_initial_status missing in blackboard",
-                self.name,
-            )
-            return Status.FAILURE
-        case_id = case_id_obj
-        if not isinstance(case_id, str):
-            status_context = _as_id(initial_status.context)
-            case_id = status_context if status_context is not None else None
-        if case_id is None:
-            self.logger.error("%s: case_id not available", self.name)
-            return Status.FAILURE
-
-        self.blackboard.new_case_participant = VultronParticipant(
-            attributed_to=self.actor_id,
-            context=case_id,
-            case_roles=_effective_case_roles(self.actor_config),
-            participant_statuses=[initial_status],
-        )
-        return Status.SUCCESS
-
-
-class AttachOwnerParticipantToCaseNode(DataLayerAction):
-    """Persist and attach staged owner participant to the case."""
-
-    def __init__(self, name: str | None = None) -> None:
-        super().__init__(name=name or self.__class__.__name__)
-
-    def setup(self, **kwargs: Any) -> None:
-        super().setup(**kwargs)
-        self.blackboard.register_key(
-            key="case_id", access=py_trees.common.Access.READ
-        )
-        self.blackboard.register_key(
-            key="new_case_participant",
-            access=py_trees.common.Access.READ,
-        )
-        self.blackboard.register_key(
-            key="participant_case", access=py_trees.common.Access.WRITE
-        )
-
-    def update(self) -> Status:
-        if self.datalayer is None or self.actor_id is None:
-            self.logger.error(
-                "%s: DataLayer or actor_id not available",
-                self.name,
-            )
-            return Status.FAILURE
-        try:
-            case_id_obj = self.blackboard.get("case_id")
-        except KeyError:
-            case_id_obj = None
-        participant = self.blackboard.get("new_case_participant")
-        if not isinstance(participant, VultronParticipant):
-            self.logger.error(
-                "%s: case_id/new_case_participant missing in blackboard",
-                self.name,
-            )
-            return Status.FAILURE
-        case_id = case_id_obj
-        if not isinstance(case_id, str):
-            case_id = _as_id(participant.context)
-        if case_id is None:
-            self.logger.error("%s: case_id not available", self.name)
-            return Status.FAILURE
-
-        stored_case = _create_and_attach_participant(
-            self.datalayer,
-            participant,
-            case_id,
-            self.actor_id,
-            self.logger,
-        )
-        if stored_case is None:
-            self.logger.error(
-                "%s: Case %s not found in DataLayer",
-                self.name,
-                case_id,
-            )
-            return Status.FAILURE
-
-        self.blackboard.participant_case = stored_case
-        return Status.SUCCESS
-
-
-class PersistOwnerCaseNode(DataLayerAction):
-    """Persist the updated case after owner participant attachment."""
-
-    def __init__(self, name: str | None = None) -> None:
-        super().__init__(name=name or self.__class__.__name__)
-
-    def setup(self, **kwargs: Any) -> None:
-        super().setup(**kwargs)
-        self.blackboard.register_key(
-            key="participant_case", access=py_trees.common.Access.READ
-        )
-
-    def update(self) -> Status:
-        if self.datalayer is None:
-            self.logger.error("%s: DataLayer not available", self.name)
-            return Status.FAILURE
-        stored_case = self.blackboard.get("participant_case")
-        if not is_case_model(stored_case):
-            self.logger.error(
-                "%s: participant_case missing in blackboard",
-                self.name,
-            )
-            return Status.FAILURE
-        self.datalayer.save(stored_case)
-        return Status.SUCCESS
-
-
-class ShouldAdvanceOwnerToAcceptedNode(py_trees.behaviour.Behaviour):
-    """Condition leaf for owner RM advancement branch selection."""
-
-    def __init__(
-        self, advance_to_accepted: bool, name: str | None = None
-    ) -> None:
-        super().__init__(name=name or self.__class__.__name__)
-        self._advance_to_accepted = advance_to_accepted
-
-    def update(self) -> Status:
-        return Status.SUCCESS if self._advance_to_accepted else Status.FAILURE
-
-
-class AdvanceOwnerRmToAcceptedNode(DataLayerAction):
-    """Advance owner RM to ACCEPTED when case creation means engagement."""
-
-    def __init__(self, name: str | None = None) -> None:
-        super().__init__(name=name or self.__class__.__name__)
-
-    def setup(self, **kwargs: Any) -> None:
-        super().setup(**kwargs)
-        self.blackboard.register_key(
-            key="case_id", access=py_trees.common.Access.READ
-        )
-        self.blackboard.register_key(
-            key="participant_case", access=py_trees.common.Access.READ
-        )
-
-    def update(self) -> Status:
-        if self.datalayer is None or self.actor_id is None:
-            self.logger.error(
-                "%s: DataLayer or actor_id not available",
-                self.name,
-            )
-            return Status.FAILURE
-        try:
-            case_id_obj = self.blackboard.get("case_id")
-        except KeyError:
-            case_id_obj = None
-        case_id = case_id_obj if isinstance(case_id_obj, str) else None
-        if case_id is None:
-            stored_case = self.blackboard.get("participant_case")
-            case_id = stored_case.id_ if is_case_model(stored_case) else None
-        if case_id is None:
-            self.logger.error("%s: case_id not available", self.name)
-            return Status.FAILURE
-
-        advanced = update_participant_rm_state(
-            case_id,
-            self.actor_id,
-            RM.ACCEPTED,
-            self.datalayer,
-        )
-        if advanced:
-            self.logger.info(
-                "Owner RM: VALID → ACCEPTED for actor '%s' in case '%s' "
-                "(case creation = case engagement)",
-                self.actor_id,
-                case_id,
-            )
-        else:
-            self.logger.warning(
-                "%s: Could not advance owner RM to ACCEPTED for actor '%s'"
-                " in case '%s'",
-                self.name,
-                self.actor_id,
-                case_id,
-            )
-        return Status.SUCCESS
-
-
-class RecordOwnerJoinedEventNode(DataLayerAction):
-    """Record owner_joined event and persist the case update.
-
-    Per notes/bt-integration.md, all protocol-significant behavior MUST be in
-    the BT.  This node records the case history event for the owner's initial
-    participation, analogous to ``RecordParticipantAddedEventNode`` in the
-    non-owner path (CM-02-008).
-    """
-
-    def __init__(self, name: str | None = None) -> None:
-        super().__init__(name=name or self.__class__.__name__)
-
-    def setup(self, **kwargs: Any) -> None:
-        super().setup(**kwargs)
-        self.blackboard.register_key(
-            key="participant_case",
-            access=py_trees.common.Access.READ,
-        )
-        self.blackboard.register_key(
-            key="new_case_participant",
-            access=py_trees.common.Access.READ,
-        )
-
-    def update(self) -> Status:
-        if self.datalayer is None:
-            self.logger.error("%s: DataLayer not available", self.name)
-            return Status.FAILURE
-
-        stored_case = self.blackboard.get("participant_case")
-        participant = self.blackboard.get("new_case_participant")
-        if not is_case_model(stored_case) or not isinstance(
-            participant, VultronParticipant
-        ):
-            self.logger.error(
-                "%s: participant_case/new_case_participant missing in"
-                " blackboard",
-                self.name,
-            )
-            return Status.FAILURE
-
-        stored_case.record_event(participant.id_, "owner_joined")
-        self.datalayer.save(stored_case)
-        return Status.SUCCESS
-
-
-class CreateCaseOwnerParticipant(py_trees.composites.Sequence):
-    """
-    Composed subtree that creates and attaches the case-owner participant.
-
-    Per specs/case-management.yaml CM-02-008, BTND-05-002, and BTND-07-001.
-    """
-
-    def __init__(
-        self,
-        actor_config: ActorConfig | None = None,
-        report_id: str | None = None,
-        case_obj: VultronCase | None = None,
-        advance_to_accepted: bool = False,
-        initial_rm_state: RM = RM.VALID,
-        name: str | None = None,
-    ):
-        super().__init__(
-            name=name or self.__class__.__name__,
-            memory=False,
-            children=[
-                ResolveOwnerInitialStatusNode(
-                    report_id=report_id,
-                    case_obj=case_obj,
-                    initial_rm_state=initial_rm_state,
-                ),
-                CreateOwnerParticipantNode(actor_config=actor_config),
-                AttachOwnerParticipantToCaseNode(),
-                PersistOwnerCaseNode(),
-                RecordOwnerJoinedEventNode(),
-                py_trees.composites.Selector(
-                    name="AdvanceOwnerRmIfConfigured",
-                    memory=False,
-                    children=[
-                        py_trees.composites.Sequence(
-                            name="AdvanceOwnerRmBranch",
-                            memory=False,
-                            children=[
-                                ShouldAdvanceOwnerToAcceptedNode(
-                                    advance_to_accepted=advance_to_accepted
-                                ),
-                                AdvanceOwnerRmToAcceptedNode(),
-                            ],
-                        ),
-                        py_trees.behaviours.Success(name="SkipAdvanceOwnerRm"),
-                    ],
-                ),
-            ],
-        )
-
-
-class ResolveParticipantAcceptedStatusNode(DataLayerAction):
-    """Resolve or create report-phase RM.ACCEPTED status for the participant."""
-
-    def __init__(
-        self,
-        participant_actor_id: str,
-        report_id: str | None = None,
-        name: str | None = None,
-    ) -> None:
-        super().__init__(name=name or self.__class__.__name__)
-        self.participant_actor_id = participant_actor_id
-        self.report_id = report_id
-
-    def setup(self, **kwargs: Any) -> None:
-        super().setup(**kwargs)
-        self.blackboard.register_key(
-            key="participant_accepted_status",
-            access=py_trees.common.Access.WRITE,
-        )
-
-    def update(self) -> Status:
-        if self.datalayer is None:
-            self.logger.error("%s: DataLayer not available", self.name)
-            return Status.FAILURE
-
-        self.blackboard.participant_accepted_status = (
-            _get_or_create_accepted_status(
-                self.datalayer,
-                self.participant_actor_id,
-                self.report_id,
-                self.name,
-                self.logger,
-            )
-        )
-        return Status.SUCCESS
-
-
-class CreateParticipantNode(DataLayerAction):
-    """Create an in-memory VultronParticipant and store it on the blackboard."""
-
-    def __init__(
-        self,
-        participant_actor_id: str,
-        roles: list[CVDRole],
-        name: str | None = None,
-    ) -> None:
-        super().__init__(name=name or self.__class__.__name__)
-        self.participant_actor_id = participant_actor_id
-        self.roles = roles
-
-    def setup(self, **kwargs: Any) -> None:
-        super().setup(**kwargs)
-        self.blackboard.register_key(
-            key="case_id", access=py_trees.common.Access.READ
-        )
-        self.blackboard.register_key(
-            key="participant_accepted_status",
-            access=py_trees.common.Access.READ,
-        )
-        self.blackboard.register_key(
-            key="new_case_participant",
-            access=py_trees.common.Access.WRITE,
-        )
-        self.blackboard.register_key(
-            key="new_participant_id",
-            access=py_trees.common.Access.WRITE,
-        )
-
-    def update(self) -> Status:
-        case_id = self.blackboard.get("case_id")
-        if not isinstance(case_id, str):
-            self.logger.error("%s: case_id not found in blackboard", self.name)
-            return Status.FAILURE
-
-        accepted_status = self.blackboard.get("participant_accepted_status")
-        if accepted_status is not None and not isinstance(
-            accepted_status, ParticipantStatus
-        ):
-            self.logger.error(
-                "%s: participant_accepted_status has invalid type",
-                self.name,
-            )
-            return Status.FAILURE
-
-        participant = VultronParticipant(
-            attributed_to=self.participant_actor_id,
-            context=case_id,
-            case_roles=self.roles,
-            participant_statuses=(
-                [accepted_status] if accepted_status is not None else []
-            ),
-        )
-        self.blackboard.new_case_participant = participant
-        self.blackboard.new_participant_id = participant.id_
-        return Status.SUCCESS
-
-
-class AttachParticipantToCaseNode(DataLayerAction):
-    """Attach the participant to case surfaces and persist the participant row."""
-
-    def __init__(self, participant_actor_id: str, name: str | None = None):
-        super().__init__(name=name or self.__class__.__name__)
-        self.participant_actor_id = participant_actor_id
-
-    def setup(self, **kwargs: Any) -> None:
-        super().setup(**kwargs)
-        self.blackboard.register_key(
-            key="case_id", access=py_trees.common.Access.READ
-        )
-        self.blackboard.register_key(
-            key="new_case_participant",
-            access=py_trees.common.Access.READ,
-        )
-        self.blackboard.register_key(
-            key="participant_case",
-            access=py_trees.common.Access.WRITE,
-        )
-
-    def update(self) -> Status:
-        if self.datalayer is None:
-            self.logger.error("%s: DataLayer not available", self.name)
-            return Status.FAILURE
-
-        case_id = self.blackboard.get("case_id")
-        participant = self.blackboard.get("new_case_participant")
-        if not isinstance(case_id, str):
-            self.logger.error("%s: case_id not found in blackboard", self.name)
-            return Status.FAILURE
-        if not isinstance(participant, VultronParticipant):
-            self.logger.error(
-                "%s: new_case_participant not found in blackboard",
-                self.name,
-            )
-            return Status.FAILURE
-
-        stored_case = _create_and_attach_participant(
-            self.datalayer,
-            participant,
-            case_id,
-            self.participant_actor_id,
-            self.logger,
-        )
-        if stored_case is None:
-            self.logger.error(
-                "%s: Case %s not found in DataLayer",
-                self.name,
-                case_id,
-            )
-            return Status.FAILURE
-
-        self.blackboard.participant_case = stored_case
-        return Status.SUCCESS
-
-
-class RecordParticipantAddedEventNode(DataLayerAction):
-    """Record participant_added event and persist case updates."""
-
-    def __init__(self, name: str | None = None) -> None:
-        super().__init__(name=name or self.__class__.__name__)
-
-    def setup(self, **kwargs: Any) -> None:
-        super().setup(**kwargs)
-        self.blackboard.register_key(
-            key="participant_case",
-            access=py_trees.common.Access.READ,
-        )
-        self.blackboard.register_key(
-            key="new_participant_id",
-            access=py_trees.common.Access.READ,
-        )
-
-    def update(self) -> Status:
-        if self.datalayer is None:
-            self.logger.error("%s: DataLayer not available", self.name)
-            return Status.FAILURE
-
-        stored_case = self.blackboard.get("participant_case")
-        participant_id = self.blackboard.get("new_participant_id")
-        if not is_case_model(stored_case) or not isinstance(
-            participant_id, str
-        ):
-            self.logger.error(
-                "%s: participant_case/new_participant_id missing in blackboard",
-                self.name,
-            )
-            return Status.FAILURE
-
-        stored_case.record_event(participant_id, "participant_added")
-        self.datalayer.save(stored_case)
-        return Status.SUCCESS
-
-
-class CaseHasActiveEmbargoNode(DataLayerAction):
-    """Condition node: SUCCESS when the case has an active embargo."""
-
-    def __init__(self, name: str | None = None) -> None:
-        super().__init__(name=name or self.__class__.__name__)
-
-    def setup(self, **kwargs: Any) -> None:
-        super().setup(**kwargs)
-        self.blackboard.register_key(
-            key="participant_case",
-            access=py_trees.common.Access.READ,
-        )
-
-    def update(self) -> Status:
-        stored_case = self.blackboard.get("participant_case")
-        if not is_case_model(stored_case):
-            self.logger.error(
-                "%s: participant_case missing in blackboard", self.name
-            )
-            return Status.FAILURE
-        return (
-            Status.SUCCESS
-            if _as_id(stored_case.active_embargo) is not None
-            else Status.FAILURE
-        )
-
-
-class CaseHasNoActiveEmbargoNode(DataLayerAction):
-    """Condition node: SUCCESS when no active embargo exists for this case."""
-
-    def __init__(self, name: str | None = None) -> None:
-        super().__init__(name=name or self.__class__.__name__)
-
-    def setup(self, **kwargs: Any) -> None:
-        super().setup(**kwargs)
-        self.blackboard.register_key(
-            key="participant_case",
-            access=py_trees.common.Access.READ,
-        )
-
-    def update(self) -> Status:
-        stored_case = self.blackboard.get("participant_case")
-        if not is_case_model(stored_case):
-            self.logger.error(
-                "%s: participant_case missing in blackboard", self.name
-            )
-            return Status.FAILURE
-        return (
-            Status.SUCCESS
-            if _as_id(stored_case.active_embargo) is None
-            else Status.FAILURE
-        )
-
-
-class SeedParticipantAsSignatoryNode(DataLayerAction):
-    """Seed the new participant as SIGNATORY when an embargo is active."""
-
-    def __init__(self, participant_actor_id: str, name: str | None = None):
-        super().__init__(name=name or self.__class__.__name__)
-        self.participant_actor_id = participant_actor_id
-
-    def setup(self, **kwargs: Any) -> None:
-        super().setup(**kwargs)
-        self.blackboard.register_key(
-            key="participant_case",
-            access=py_trees.common.Access.READ,
-        )
-        self.blackboard.register_key(
-            key="new_case_participant",
-            access=py_trees.common.Access.READ,
-        )
-
-    def update(self) -> Status:
-        if self.datalayer is None:
-            self.logger.error("%s: DataLayer not available", self.name)
-            return Status.FAILURE
-
-        stored_case = self.blackboard.get("participant_case")
-        participant = self.blackboard.get("new_case_participant")
-        if not is_case_model(stored_case) or not isinstance(
-            participant, VultronParticipant
-        ):
-            self.logger.error(
-                "%s: participant_case/new_case_participant missing in blackboard",
-                self.name,
-            )
-            return Status.FAILURE
-
-        active_embargo_id = _as_id(stored_case.active_embargo)
-        if active_embargo_id is None:
-            self.logger.error(
-                "%s: cannot seed SIGNATORY without active embargo",
-                self.name,
-            )
-            return Status.FAILURE
-
-        participant.embargo_consent_state = PEC.SIGNATORY
-        if active_embargo_id not in participant.accepted_embargo_ids:
-            participant.accepted_embargo_ids.append(active_embargo_id)
-        self.datalayer.save(participant)
-        self.logger.info(
-            "Seeded participant '%s' (actor '%s') as SIGNATORY"
-            " for active embargo in case '%s' (CM-14-005)",
-            participant.id_,
-            self.participant_actor_id,
-            stored_case.id_,
-        )
-        return Status.SUCCESS
-
-
-class SeedParticipantAsSignatoryIfEmbargoActiveNode(
-    py_trees.composites.Selector
-):
-    """Conditional subtree for CM-14-005 signatory seeding behavior."""
-
-    def __init__(self, participant_actor_id: str, name: str | None = None):
-        super().__init__(
-            name=name or self.__class__.__name__,
-            memory=False,
-            children=[
-                py_trees.composites.Sequence(
-                    name="SeedWhenActiveEmbargo",
-                    memory=False,
-                    children=[
-                        CaseHasActiveEmbargoNode(),
-                        SeedParticipantAsSignatoryNode(
-                            participant_actor_id=participant_actor_id
-                        ),
-                    ],
-                ),
-                CaseHasNoActiveEmbargoNode(),
-            ],
-        )
-
-
-class QueueAddParticipantNotificationNode(DataLayerAction):
-    """Queue Add(CaseParticipant) outbox notification for the sender actor."""
-
-    def __init__(self, participant_actor_id: str, name: str | None = None):
-        super().__init__(name=name or self.__class__.__name__)
-        self.participant_actor_id = participant_actor_id
-
-    def setup(self, **kwargs: Any) -> None:
-        super().setup(**kwargs)
-        self.blackboard.register_key(
-            key="case_id", access=py_trees.common.Access.READ
-        )
-        self.blackboard.register_key(
-            key="new_participant_id",
-            access=py_trees.common.Access.READ,
-        )
-
-    def update(self) -> Status:
-        if self.datalayer is None or self.actor_id is None:
-            self.logger.error(
-                "%s: DataLayer or actor_id not available", self.name
-            )
-            return Status.FAILURE
-
-        case_id = self.blackboard.get("case_id")
-        participant_id = self.blackboard.get("new_participant_id")
-        if not isinstance(case_id, str) or not isinstance(participant_id, str):
-            self.logger.error(
-                "%s: case_id/new_participant_id not found in blackboard",
-                self.name,
-            )
-            return Status.FAILURE
-
-        if not _queue_participant_add_notification(
-            self.datalayer,
-            self.name,
-            self.logger,
-            self.actor_id,
-            self.participant_actor_id,
-            participant_id,
-            case_id,
-            trigger_activity=self.trigger_activity_factory,
-        ):
-            return Status.FAILURE
-        return Status.SUCCESS
-
-
-class CreateCaseParticipantNode(py_trees.composites.Sequence):
-    """
-    Composed subtree that creates and attaches a CaseParticipant.
-
-    This decomposes participant creation into named leaf nodes so each step is
-    explicit and testable per BTND-07-001.
-    """
-
-    def __init__(
-        self,
-        actor_id: str,
-        roles: list[CVDRole],
-        report_id: str | None = None,
-        name: str | None = None,
-    ) -> None:
-        super().__init__(
-            name=name or self.__class__.__name__,
-            memory=False,
-            children=[
-                ResolveParticipantAcceptedStatusNode(
-                    participant_actor_id=actor_id,
-                    report_id=report_id,
-                ),
-                CreateParticipantNode(
-                    participant_actor_id=actor_id,
-                    roles=roles,
-                ),
-                AttachParticipantToCaseNode(participant_actor_id=actor_id),
-                RecordParticipantAddedEventNode(),
-                SeedParticipantAsSignatoryIfEmbargoActiveNode(
-                    participant_actor_id=actor_id
-                ),
-                QueueAddParticipantNotificationNode(
-                    participant_actor_id=actor_id
-                ),
-            ],
-        )
-
-
-class CreateParticipantStatusNode(DataLayerAction):
-    """Create a ParticipantStatus snapshot and append it to the participant.
-
-    BT-15-001: direct ``ParticipantStatus`` writes with an explicit
-    ``rm_state`` are protocol-significant behavior and MUST live inside a
-    BT leaf node, not directly in ``execute()``.
-
-    This node is the leaf responsible for the ParticipantStatus record
-    creation in the ``add_participant_status_trigger_bt`` tree.  It reads
-    the actor's current RM/VFD state from the DataLayer (via
-    ``resolve_participant_state_from_dl``) and creates a snapshot record
-    using the override values from the trigger request when provided.
-
-    Constructor parameters:
-        case_id: ID of the VulnerabilityCase.
-        actor_id: Actor whose state is being self-reported.
-        rm_state: RM state to record (``None`` → use current state).
-        vfd_state: CS_vfd state to record (``None`` → use current state).
-        pxa_state: CS_pxa for an optional CaseStatus snapshot
-            (``None`` → no CaseStatus attached).
-        result_out: Mutable dict populated by this node with
-            ``'status_id'`` and ``'participant_id'``; read by the
-            ``activity_builder`` closure in
-            ``SvcAddParticipantStatusUseCase``.
-
-    Writes to ``result_out``:
-        - ``status_id``: ID of the newly created ParticipantStatus.
-        - ``participant_id``: ID of the actor's CaseParticipant.
-    """
-
-    def __init__(
-        self,
-        case_id: str,
-        actor_id: str,
-        rm_state: "RM | None",
-        vfd_state: "CS_vfd | None",
-        pxa_state: "CS_pxa | None",
-        result_out: dict,
-        name: str | None = None,
-    ) -> None:
-        super().__init__(name=name or self.__class__.__name__)
-        self._case_id = case_id
-        self._actor_id = actor_id
-        self._rm_state = rm_state
-        self._vfd_state = vfd_state
-        self._pxa_state = pxa_state
-        self._result_out = result_out
-
-    def update(self) -> Status:
-        dl = self.datalayer
-        if dl is None:
-            self.logger.error("%s: DataLayer not available", self.name)
-            self.feedback_message = "DataLayer not available"
-            return Status.FAILURE
-
-        case = dl.read(self._case_id)
-        if not is_case_model(case):
-            self.logger.error(
-                "%s: Case '%s' not found in DataLayer",
-                self.name,
-                self._case_id,
-            )
-            self.feedback_message = f"Case '{self._case_id}' not found"
-            return Status.FAILURE
-
-        participant_id = case.actor_participant_index.get(self._actor_id)
-        if participant_id is None:
-            self.logger.error(
-                "%s: actor '%s' not in case '%s'",
-                self.name,
-                self._actor_id,
-                self._case_id,
-            )
-            self.feedback_message = (
-                f"Actor '{self._actor_id}' not found in"
-                f" case '{self._case_id}'"
-            )
-            return Status.FAILURE
-
-        case_status: CaseStatus | None = None
-        if self._pxa_state is not None:
-            current_em = getattr(
-                getattr(case, "current_status", None), "em_state", None
-            )
-            case_status = CaseStatus(
-                context=self._case_id,
-                attributed_to=self._actor_id,
-                em_state=current_em if current_em is not None else EM.NONE,
-                pxa_state=self._pxa_state,
-            )
-
-        current_rm, current_vfd = resolve_participant_state_from_dl(
-            dl, participant_id
-        )
-
-        status = ParticipantStatus(
-            context=self._case_id,
-            attributed_to=self._actor_id,
-            rm_state=(
-                self._rm_state if self._rm_state is not None else current_rm
-            ),
-            vfd_state=(
-                self._vfd_state if self._vfd_state is not None else current_vfd
-            ),
-            case_status=case_status,
-        )
-        try:
-            dl.create(status)
-        except ValueError:
-            dl.save(status)
-
-        participant_obj = dl.read(participant_id)
-        wire_status = dl.read(status.id_)
-        participant_statuses = (
-            getattr(participant_obj, "participant_statuses", None)
-            if participant_obj is not None
-            else None
-        )
-        if participant_statuses is not None and wire_status is not None:
-            participant_statuses.append(wire_status)
-            if participant_obj is not None:
-                dl.save(participant_obj)
-
-        self._result_out["status_id"] = status.id_
-        self._result_out["participant_id"] = participant_id
-
-        self.logger.info(
-            "%s: Created ParticipantStatus '%s' for actor '%s' in case '%s'",
-            self.name,
-            status.id_,
-            self._actor_id,
-            self._case_id,
-        )
-        return Status.SUCCESS
